ch_integration_tests: Add direct kernel boot test
Relying on the newly added KernelType enum, this patch expand the list
of valid types with "Direct", to represent the direct kernel boot type.
A new test is created to leverage this new type and make sure direct
kernel boot is properly validated through the integration test suite.

The kernel is built from the Cloud Hypervisor kernel branch.

Signed-off-by: Sebastien Boeuf <sebastien.boeuf@intel.com>